# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Opportunity Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsByStatus } from './dashboardStatsByStatus';
import type { DashboardStatsByType } from './dashboardStatsByType';

export interface DashboardStats {
  total: number;
  closingSoon: number;
  byStatus: DashboardStatsByStatus;
  byType: DashboardStatsByType;
}