# lib/api-zod/src/generated/types/opportunity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Opportunity Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunityStatus } from './opportunityStatus';
import type { OpportunityType } from './opportunityType';

export interface Opportunity {
  id: number;
  url: string;
  title: string;
  type: OpportunityType;
  status: OpportunityStatus;
  /**
     * ISO date string YYYY-MM-DD
     * @nullable
     */
  deadline?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  keyActionSteps?: string | null;
  createdAt: string;
  taskCount?: number;
  completedTaskCount?: number;
}