# lib/api-zod/src/generated/types/opportunityInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Opportunity Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunityInputStatus } from './opportunityInputStatus';
import type { OpportunityInputType } from './opportunityInputType';

export interface OpportunityInput {
  url: string;
  /** @minLength 1 */
  title: string;
  type: OpportunityInputType;
  status: OpportunityInputStatus;
  deadline?: string;
  summary?: string;
  keyActionSteps?: string;
}