# lib/api-zod/src/generated/types/scrapedData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Opportunity Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScrapedData {
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  keyActionSteps?: string | null;
  scrapeSuccess?: boolean;
}