# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Opportunity Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: number;
  opportunityId: number;
  title: string;
  completed: boolean;
  createdAt: string;
}